switched resize and fullscreen here to component as well
