+++ frontend/src/components/game/LoseScreen.js
@@ -1,7 +1,11 @@
 import React, { useState } from 'react';
+import { useTranslation } from "react-i18next";
+import { useNavigate } from "react-router-dom";
 
-const LoseScreen = ({ t, navigate, GameCanvas, backgroundImage, WelcomeButtonStyle, BackButton }) => {
+const LoseScreen = ({ GameCanvas, backgroundImage, WelcomeButtonStyle, BackButton }) => {
     const [gameStarted, setGameStarted] = useState(false);
+    const { t } = useTranslation();
+    const navigate = useNavigate();
 
     const handleButtonClick = () => {
         setGameStarted(true);
